- Add comparisons for Java EnumSet and Scala Set - Add Map comparisons
Results
-------

[info] Benchmark                                  Mode  Cnt   Score   Error  Units
[info] MapComparisons.enumeratumScalaMapGet       avgt   30   9.830 ± 0.207  ns/op
[info] MapComparisons.jEnumEnumMapGet             avgt   30   4.745 ± 0.685  ns/op
[info] MapComparisons.jEnumScalaMapGet            avgt   30  12.204 ± 0.186  ns/op
[info] SetComparisons.enumeratumScalaSetContains  avgt   30   7.670 ± 1.333  ns/op
[info] SetComparisons.jEnumEnumSetContains        avgt   30   4.391 ± 0.077  ns/op
[info] SetComparisons.jEnumScalaSetContains       avgt   30   8.906 ± 0.235  ns/op

Discussion
----------

There is no denying that Java  EnumMap and EnumSet are fast. Given that JMH supposedly has
an over head of 2-3ns, they both offer roughly 2-3ns/op, which might just be noise !

On the other hand, plain old Scala Set and Map aren't bad either, coming in at < 10ns/op,
which is maybe 4-6ns/op accounting for overhead and noise.

So the current state of things is that Java EnumMap and EnumSet are roughly 2x faster than
plain old Scala Set and Maps w/ Enumeratum, but being that we're still in the sub-10ns
range, it seems reasonable to say that performance here won't be a bottleneck except in super
edge cases.