Upgrade sbt-doctest
The version of sbt-doctest used by enumeratum is not compatible with
scala 2.12.11 or 2.12.12 because scala broke binary compatibility in one
of the apis called by the plugin. Later versions of the plugin use an
entirely different mechanism that is unaffected by the broken binary
compatibility. This change will likely be necessary for enumeratum to
work with sbt 1.4.0 because it will use scala 2.12.12.