Make findValues less strict about types so higher-kinded types work
Some code of the form `extends StringEnum[Entry[?]]` that worked in
scala 2 no longer works in scala 3, as existential types were removed.
The simplest way to get the types to work is to change the `?` to `Any`.
But then `findValues` can't process all the enum entries because they
don't extend `Entry[Any]`, they extend `Entry[X]` for some type `X`.

The macro was previously quite strict about types. It would get the
`MirroredElemTypes` from a `Mirror.SumOf[A]`, and then while processing
the types, it would check each element to make sure it extended `A`. But
the elements are from the mirror, which just has the things that extend
`A`, so of course they extend `A`. It would also check while looking for
the constructor to pull the value from that the constructor type
extended `A`. I updated this check to unwrap both types if either was an
`AppliedType`. I.e. instead of checking that `Entry[Int] <:< Entry[Any]`
it checks that `Entry <:< Entry`.

This is a slight change from the scala 2 macro behavior, but it allows
higher kinded type parameters to be used in enums without needing
existential types.